requirements.txt
email-validator>=2.3.0
flask>=3.1.2
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
pytz>=2025.2
requests>=2.32.5
sqlalchemy>=2.0.43
tigeropen>=3.4.6
werkzeug>=3.1.3
websockets
anthropic
tenacity
